+++ BinWeevils.Protocol/XmlMessages/UserRecord.cs
@@ -5,6 +5,6 @@ namespace BinWeevils.Protocol.XmlMessages
     [XmlCls("user")]
     public partial struct UserRecord
     {
-        [XmlField("id")] public int m_id;
+        [XmlField("id")] public ulong m_id;
     }
 }
